# Speed up `optimize_projection`