## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

- Closes #.

## Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

We found the `is_projection_unnecessary` method is costly if there are many exprs in projection, and is a bottleneck in the `optimize_projection` rule

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

Speed up the `is_projection_necessary function`
1. Add the check for the `exprs` length with the `input.schema()`
2. Optimize the way of checking if the projection can be removed.

### Benchmark result (~1000x)
```
projection_unnecessary_comparison/is_projection_unnecessary_v2
                        time:   [3.1893 µs 3.2122 µs 3.2391 µs]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  3 (3.00%) high severe
projection_unnecessary_comparison/is_projection_unnecessary
                        time:   [4.0953 ms 4.1299 ms 4.1644 ms]
Found 25 outliers among 100 measurements (25.00%)
  10 (10.00%) low severe
  12 (12.00%) high mild
  3 (3.00%) high severe
```

## Are these changes tested?

Yes, by existing tests

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
4. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->