@xudong963 2025-04-21
> 🤖 `./gh_compare_branch.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_branch.sh) Running Linux aal-dev 6.8.0-1016-gcp #18-Ubuntu SMP Fri Oct 4 22:16:29 UTC 2024 x86_64 x86_64 x86_64 GNU/Linux Comparing speed_up_optimize_projection ([7f7e4c3](https://github.com/apache/datafusion/commit/7f7e4c3affd09a2d15bb27738dcf0aea2cc4d396)) to [375189c](https://github.com/apache/datafusion/commit/375189c90d65075e9428b6ed902fb8fafe3bfe1a) [diff](https://github.com/apache/datafusion/compare/375189c90d65075e9428b6ed902fb8fafe3bfe1a..7f7e4c3affd09a2d15bb27738dcf0aea2cc4d396) Benchmarks: tpch_mem clickbench_partitioned clickbench_extended Results will be posted here when complete

@alamb How long does it usually take to generate results? 🤔